feat: implement auto-system-config-generator with unified pipeline for YAML generation
- Introduced a new script `generate_system_config.py` to facilitate the generation of Autoware System Designer YAML configurations from ROS 2 launch files.
- Added support for a unified pipeline that includes parsing launch files, capturing live ROS 2 graph snapshots, and merging topics into node records.
- Created a structured output directory for generated files, including system, module, and node YAML configurations.
- Enhanced the README.md to provide comprehensive usage instructions and document new features.

These changes significantly improve the usability and flexibility of the auto-system-config-generator, streamlining the process of generating system configurations for Autoware.
