+++ tools/system-config-generator/config/component_map.yaml
