+++ tools/system-config-generator/launch_unifier/templates/launch_generated.jinja2
