+++ tools/system-config-generator/pipeline/emitter/__init__.py
@@ -1,0 +1,10 @@
+from .module import _collect_all_pub_sub, emit_module_yaml_from_tree
+from .parameter import emit_parameter_set_yaml
+from .system import emit_system_yaml_from_tree
+
+__all__ = [
+    "_collect_all_pub_sub",
+    "emit_module_yaml_from_tree",
+    "emit_parameter_set_yaml",
+    "emit_system_yaml_from_tree",
+]
