+++ tools/system-config-generator/pipeline/graph_parser.py
