+++ tools/system-config-generator/pipeline/launch_parser.py
