+++ tools/system-config-generator/pipeline/namespace_tree.py
